app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
redis>=5.0.0
requests>=2.31.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0